# Added caching of children to avoid traversing the file system on every file request