@jkeiser 2015-08-03
Hmm, we'd been trying to avoid caching children because it was causing unhappiness--we were trying to keep a mostly-immediate API and optimize the algorithms so they only called `.children` once. But, we've noticed the slowness too, and good lord this has to get fixed soon, so I'd take this if an algorithm fix didn't appear.

Do you know who the callers are that are causing this? It sounds like it happens on a simple directory list?