@tpodom 2015-08-04
Yeah the caching is a good point.  In fact that's what's causing the test failures one of the cookbook deletes isn't getting removed from the caches causing later tests to fail.

As far as what's causing the extra iterations it's two things:

1.  In ```Chef::ChefFS::ChefFSDataStore.with_entry``` it calls ```to_chef_fs_path``` which calls ```get_single_cookbook_version``` which results in a call to ```Chef::ChefFS::FileSystem.resolve_path``` which in turn iterates each path segment asking for the child. Then after ```to_chef_fs_path``` finishes ```with_entry``` turns around and calls ```Chef::ChefFS::FileSystem.resolve_path``` again which again iterates the children to retrieve a specific child.  That results in 2 iterations of all of the cookbooks on a single GET

2.  From chef-client's perspective it's pretty much doing 1 or 2 passes through for every LIST or GET because LIST will at least call 

From looking through this more I'd say the best solution is to avoid calling children on every ```child(name)``` call as proposed by #2007.