@jkeiser 2015-08-04
@tpodom thanks tons for the info! I think you missed the end of the sentence on your second point? I'm not clear what that one is about.

I've started updating #2007, but I'm not sure about your specific solution there--`child` is required to return something in all cases, though it doesn't *have* to be a cookbook dir, it could be a Nonexistent entry.