@jkeiser 2015-08-04
Regarding number 1: `child(name)` is supposed to be a costless operation. Is that not the case for one of these? That would totally be the cause, as it is baked into all our assumptions.