@tpodom 2015-08-04
@jkeiser Right now for every call to ```resolve_path``` it's iterating the entire cookbook directory structure. ```resolve_path`` is calling ```child(name)``` on ```MultiplexedDir``` which is handled by ```BaseFSDir::child(name)``` which results in calling ```MultiplexedDir::children```.  ```MultiplexedDir::children``` is then triggering calls to ```children``` on all of the structure which eventually gets to ```ChefRepositoryFileSystemCookbooksDir::children``` which walks all of the cookbook directories and calls ```ChefRepositoryFileSystemCookbookDir::children``` which walks each individual cookbook.  

My point about #2 can mostly be ignored, I was trying trying to say every time chef-client calls LIST or GET it's iterating the entire cookbook directory structure at least once because ```resolve_path``` gets called.  With the way ```ChefFSDataStore::with_entry``` is written any time that is called ```resolve_path``` is called twice resulting in two iterations of the cookbook directory structure.  In one of the cases of LIST it doesn't call ```with_entry``` but it does call ```get_single_cookbook_version``` which calls ```resolve_path``` triggering the directory walking.  All of that wouldn't really matter if the ```child(name)``` call wasn't expensive for ```MultiplexedDir``` and ```ChefRepositoryFileSystemCookbooksDir``` classes.

On #2007 I was just trying to come up with an example for implementing ```child(name)``` but you are right I was failing to return the ```NonexistentFSObject```.  I also had missed at first that you had modified the base ```FileSystemEntry``` to prevent the ```ChefRepositoryFileSystem*``` classes from iterating children on ```child(name)```.  However, I wonder if the implementation of the ```FileSystemEntry::child(name)``` should ask if the child is valid to allow for some of the filtering logic that's performed in the ```children``` functions of the subclasses?