@jkeiser 2015-08-04
@tpodom does this happen only when synchronizing a cookbook for the first time, or is it also slow the second time? I've got me a repro and I'm trying to figure out just what's *supposed* to be slow and what isn't.