@tpodom 2015-08-04
It happens every time chef-client does a LIST or GET on any cookbook (and I supposed on any file contained within the cookbook too but I haven't verified that).  It appears that chef-client does a LIST on the cookbook then performs a GET so that combination results in 3 complete directory traversals just to retrieve the cookbook metadata.  It then does that for every cookbook in the run list so if there are 10 cookbooks it iterates the complete cookbook directory structure 30 times.