@jkeiser 2015-08-04
@tpodom OK cool. Was just checking. So that means if you run `chef-client` multiple times in a row, it will continue to repro, if I read it correctly.