@tpodom 2015-08-04
@jkeiser correct, every chef-client run takes > 30 minutes for us to synchronize cookbooks.