@jkeiser 2015-08-04
I will note that with a rather large run list (which is nonetheless stored entirely locally), it takes me upwards of 15 minutes to sync the first time, and 1.8 seconds to sync the second time. That's why I ask: it feels like the deeper problem exists on the initial sync, even though 1.8 seconds is nothing to sneeze at.

My repro:

1. `mkdir repo`
2. `cd repo`
3. Put this in `cookbooks/blah/metadata.rb`:
```ruby
name 'blah'
version "1.5.1"
description "x"
depends "java"
depends "jenkins"
depends "windows"
depends "yum"
depends "mysql"
depends "postgresql"
depends "apache2"
depends "docker"
depends "newrelic"
depends "bacon"
depends "apt"
depends "build-essential"
depends "chef-client"
```
4. Make an empty `cookbooks/blah/recipes/default.rb`
5. Put this in `Policyfile.rb`:
```ruby
name "me"
default_source :community
run_list "blah"
cookbook "blah", path: "cookbooks/blah"
```
6. Run `chef install`
7. Run `chef export repo`
8. Put this in `repo/.chef/knife.rb`:
```ruby
use_policyfile true
policy_document_native_api false
deployment_group 'me-local'
versioned_cookbooks true
```
9. `cd repo`
10. `chef-client -z`

And that will repro.