@tpodom 2015-08-04
@jkeiser There could also be something different between your repo setup and ours.  We're still using the old style repo where we're storing all of the cookbooks, environments, roles, data bags in a single git repo.  I have a test rb file that I wrote that I can upload when I get home that just mocks up two back to back get requests on the same cookbook.