@jkeiser 2015-08-04
That would be great!  And I really really appreciate your sleuthing tracking this down.

I'm using an old style repo as well (`chef export` creates an old style repository with your chosen cookbooks in it). It's worth running the repro to see if we're seeing the same thing, IMO.