@tpodom 2015-08-04
@jkeiser OK I did just try out your example and the first sync was definitely slow but subsequent runs were fast.  One difference is that your example is using versioned cookbooks so I'm not sure if that is changing any different behavior.