@jkeiser 2015-08-04
@tpodom it's possible but probably isn't. I think though, that the first-sync case is the one that we're trying to optimize here. I see a huge difference with and without patch.