@jkeiser 2015-08-04
Docker in particular is the worst offender I've seen. I'm going to try loading just that and see what I see.