@tpodom 2015-08-05
@jkeiser the test script i wrote that's setting up the data store isn't working correctly with the versioned cookbooks set up from the ```chef export```.  I'll see if I can get a different test working.